Added support for retrieving albums.
So far, album art is unable to be retrieved, but I'll implemented that
later on. The "Song Details" dialog was also entirely revamped with a
bunch of new data. The application is now properly signed with a proper
keystore. The application also supports dark mode with Android 10's Dark
Mode setting.

Signed-off-by: Kamran Mackey <kamranm1200@gmail.com>